Fjernet enkelt vedtak tiltak
