+++ apps/testnorge-arena/src/main/java/no/nav/registre/testnorge/arena/consumer/rs/TiltakSyntConsumer.java
@@ -14,32 +14,12 @@ public class TiltakSyntConsumer {
 
     private final WebClient webClient;
 
-    private static final String ARENA_TILTAKSDELTAKELSE_URL = "/v1/arena/tiltak/deltakelse";
-    private static final String ARENA_TILTAKSPENGER_URL = "/v1/arena/tiltak/basi";
-    private static final String ARENA_BARNETILLEGG_URL = "/v1/arena/tiltak/btil";
-    private static final String ARENA_DELTAKERSTATUS_URL = "/v1/arena/tiltak/deltakerstatus";
     private static final String ARENA_TILTAKSAKTIVITET_URL = "/v1/arena/tiltak/aktivitet";
 
     public TiltakSyntConsumer(
             @Value("${synt-arena-tiltak.rest-api.url}") String arenaTiltakServerUrl
     ) {
         this.webClient = WebClient.builder().baseUrl(arenaTiltakServerUrl).build();
-    }
-
-    public List<NyttVedtakTiltak> opprettTiltaksdeltakelse(List<SyntRequest> syntRequest) {
-        return new PostSyntTiltakRequestCommand(webClient, syntRequest, ARENA_TILTAKSDELTAKELSE_URL).call();
-    }
-
-    public List<NyttVedtakTiltak> opprettTiltakspenger(List<SyntRequest> syntRequest) {
-        return new PostSyntTiltakRequestCommand(webClient, syntRequest, ARENA_TILTAKSPENGER_URL).call();
-    }
-
-    public List<NyttVedtakTiltak> opprettBarnetillegg(List<SyntRequest> syntRequest) {
-        return new PostSyntTiltakRequestCommand(webClient, syntRequest, ARENA_BARNETILLEGG_URL).call();
-    }
-
-    public List<NyttVedtakTiltak> opprettDeltakerstatus(List<SyntRequest> syntRequest) {
-        return new PostSyntTiltakRequestCommand(webClient, syntRequest, ARENA_DELTAKERSTATUS_URL).call();
     }
 
     public List<NyttVedtakTiltak> opprettTiltaksaktivitet(List<SyntRequest> syntRequest) {
