+++ apps/testnorge-arena/src/main/java/no/nav/registre/testnorge/arena/service/VedtakshistorikkService.java
@@ -42,7 +42,6 @@ import no.nav.registre.testnorge.arena.service.util.ArbeidssoekerUtils;
 import no.nav.registre.testnorge.arena.service.util.DatoUtils;
 import no.nav.registre.testnorge.arena.service.util.IdenterUtils;
 import no.nav.registre.testnorge.arena.service.util.ServiceUtils;
-import no.nav.registre.testnorge.arena.service.util.TiltakUtils;
 
 import no.nav.registre.testnorge.consumers.hodejegeren.response.KontoinfoResponse;
 import no.nav.registre.testnorge.domain.dto.arena.testnorge.brukere.Deltakerstatuser;
@@ -67,7 +66,6 @@ public class VedtakshistorikkService {
     private final RettighetArenaForvalterConsumer rettighetArenaForvalterConsumer;
     private final IdenterUtils identerUtils;
     private final ArbeidssoekerUtils arbeidsoekerUtils;
-    private final TiltakUtils tiltakUtils;
     private final RettighetAapService rettighetAapService;
     private final RettighetTiltakService rettighetTiltakService;
     private final DatoUtils datoUtils;
@@ -427,15 +425,15 @@ public class VedtakshistorikkService {
             arbeidsoekerUtils.opprettArbeidssoekerTiltakdeltakelse(personident, miljoe, kvalifiseringsgruppe);
 
             tiltaksdeltakelser.forEach(deltakelse -> {
-                if (tiltakUtils.harIkkeGyldigTiltakKode(deltakelse, kvalifiseringsgruppe)) {
-                    deltakelse.setTiltakKode(tiltakUtils.getGyldigTiltakKode(deltakelse, kvalifiseringsgruppe));
+                if (rettighetTiltakService.harIkkeGyldigTiltakKode(deltakelse, kvalifiseringsgruppe)) {
+                    deltakelse.setTiltakKode(rettighetTiltakService.getGyldigTiltakKode(deltakelse, kvalifiseringsgruppe));
                 }
                 deltakelse.setFodselsnr(personident);
                 deltakelse.setTiltakYtelse("J");
                 deltakelse.setTiltakVedtak("J");
             });
             tiltaksdeltakelser.forEach(deltakelse -> {
-                var tiltak = tiltakUtils.finnTiltak(personident, miljoe, deltakelse);
+                var tiltak = rettighetTiltakService.finnTiltak(personident, miljoe, deltakelse);
 
                 if (tiltak != null) {
                     deltakelse.setTiltakId(tiltak.getTiltakId());
@@ -449,7 +447,7 @@ public class VedtakshistorikkService {
             var nyeTiltaksdeltakelser = tiltaksdeltakelser.stream()
                     .filter(deltakelse -> deltakelse.getTiltakId() != null).collect(Collectors.toList());
 
-            nyeTiltaksdeltakelser = tiltakUtils.removeOverlappingTiltakVedtak(nyeTiltaksdeltakelser, historikk.getAap());
+            nyeTiltaksdeltakelser = rettighetTiltakService.removeOverlappingTiltakVedtak(nyeTiltaksdeltakelser, historikk.getAap());
 
             historikk.setTiltaksdeltakelse(nyeTiltaksdeltakelser);
         }
@@ -464,7 +462,7 @@ public class VedtakshistorikkService {
         var tiltaksdeltakelser = historikk.getTiltaksdeltakelse();
         if (tiltaksdeltakelser != null && !tiltaksdeltakelser.isEmpty()) {
             for (var deltakelse : tiltaksdeltakelser) {
-                var nyDeltakelse = tiltakUtils.getVedtakForTiltaksdeltakelseRequest(deltakelse);
+                var nyDeltakelse = rettighetTiltakService.getVedtakForTiltaksdeltakelseRequest(deltakelse);
                 var rettighetRequest = new RettighetTiltaksdeltakelseRequest(Collections.singletonList(nyDeltakelse));
 
                 rettighetRequest.setPersonident(personident);
@@ -489,11 +487,11 @@ public class VedtakshistorikkService {
             return;
         }
         for (var deltakelse : tiltaksdeltakelser) {
-            if (tiltakUtils.canSetDeltakelseTilGjennomfoeres(deltakelse, tiltak)) {
-                List<String> endringer = tiltakUtils.getFoersteEndringerDeltakerstatus(deltakelse.getTiltakAdminKode());
+            if (rettighetTiltakService.canSetDeltakelseTilGjennomfoeres(deltakelse, tiltak)) {
+                List<String> endringer = rettighetTiltakService.getFoersteEndringerDeltakerstatus(deltakelse.getTiltakAdminKode());
 
                 for (var endring : endringer) {
-                    var rettighetRequest = tiltakUtils.opprettRettighetEndreDeltakerstatusRequest(personident, miljoe,
+                    var rettighetRequest = rettighetTiltakService.opprettRettighetEndreDeltakerstatusRequest(personident, miljoe,
                             deltakelse, endring);
 
                     rettigheter.add(rettighetRequest);
@@ -519,10 +517,10 @@ public class VedtakshistorikkService {
         var tiltaksdeltakelser = vedtak.getTiltaksdeltakelse();
         if (tiltaksdeltakelser != null && !tiltaksdeltakelser.isEmpty()) {
             for (var deltakelse : tiltaksdeltakelser) {
-                if (tiltakUtils.canSetDeltakelseTilFinished(deltakelse, tiltak)) {
-                    var deltakerstatuskode = tiltakUtils.getAvsluttendeDeltakerstatus(deltakelse, tiltak).toString();
+                if (rettighetTiltakService.canSetDeltakelseTilFinished(deltakelse, tiltak)) {
+                    var deltakerstatuskode = rettighetTiltakService.getAvsluttendeDeltakerstatus(deltakelse, tiltak).toString();
 
-                    var rettighetRequest = tiltakUtils.opprettRettighetEndreDeltakerstatusRequest(personident, miljoe,
+                    var rettighetRequest = rettighetTiltakService.opprettRettighetEndreDeltakerstatusRequest(personident, miljoe,
                             deltakelse, deltakerstatuskode);
 
                     rettigheter.add(rettighetRequest);
@@ -540,10 +538,10 @@ public class VedtakshistorikkService {
         var tiltakspenger = historikk.getTiltakspenger() != null ? historikk.getTiltakspenger() : new ArrayList<NyttVedtakTiltak>();
         var tiltaksdeltakelser = historikk.getTiltaksdeltakelse();
 
-        List<NyttVedtakTiltak> nyeTiltakspenger = tiltakUtils.oppdaterVedtakslisteBasertPaaTiltaksdeltakelse(
+        List<NyttVedtakTiltak> nyeTiltakspenger = rettighetTiltakService.oppdaterVedtakslisteBasertPaaTiltaksdeltakelse(
                 tiltakspenger, tiltaksdeltakelser);
 
-        nyeTiltakspenger = tiltakUtils.removeOverlappingTiltakSequences(nyeTiltakspenger);
+        nyeTiltakspenger = rettighetTiltakService.removeOverlappingTiltakSequences(nyeTiltakspenger);
 
         if (nyeTiltakspenger != null && !nyeTiltakspenger.isEmpty()) {
             for (var vedtak : nyeTiltakspenger) {
@@ -569,10 +567,10 @@ public class VedtakshistorikkService {
 
         List<NyttVedtakTiltak> nyeBarnetillegg = new ArrayList<>();
         if (!tiltakspenger.isEmpty()) {
-            nyeBarnetillegg = tiltakUtils.oppdaterVedtakslisteBasertPaaTiltaksdeltakelse(
+            nyeBarnetillegg = rettighetTiltakService.oppdaterVedtakslisteBasertPaaTiltaksdeltakelse(
                     barnetillegg, tiltaksdeltakelser);
 
-            nyeBarnetillegg = tiltakUtils.removeOverlappingTiltakSequences(nyeBarnetillegg);
+            nyeBarnetillegg = rettighetTiltakService.removeOverlappingTiltakSequences(nyeBarnetillegg);
 
             if (nyeBarnetillegg != null && !nyeBarnetillegg.isEmpty()) {
                 for (var vedtak : nyeBarnetillegg) {
@@ -599,7 +597,7 @@ public class VedtakshistorikkService {
         if (tillegg != null && !tillegg.isEmpty() && !rettigheter.isEmpty()) {
 
             var tiltaksaktiviteter = rettighetTiltakService.getTiltaksaktivitetRettigheter(personident, miljoe, tillegg);
-            if (tiltaksaktiviteter != null && !tiltaksaktiviteter.isEmpty()){
+            if (tiltaksaktiviteter != null && !tiltaksaktiviteter.isEmpty()) {
                 rettigheter.addAll(tiltaksaktiviteter);
 
                 for (var vedtak : tillegg) {
