+++ apps/testnorge-arena/src/test/java/no/nav/registre/testnorge/arena/service/RettighetTiltakServiceTest.java
@@ -10,15 +10,12 @@ import java.util.Random;
 
 import no.nav.registre.testnorge.arena.consumer.rs.RettighetArenaForvalterConsumer;
 import no.nav.registre.testnorge.arena.consumer.rs.TiltakSyntConsumer;
-import no.nav.registre.testnorge.arena.consumer.rs.request.synt.SyntRequest;
 import no.nav.registre.testnorge.arena.consumer.rs.request.RettighetTilleggRequest;
 import no.nav.registre.testnorge.arena.consumer.rs.util.ConsumerUtils;
 import no.nav.registre.testnorge.arena.service.util.IdenterUtils;
 import no.nav.registre.testnorge.arena.service.util.ArbeidssoekerUtils;
 import no.nav.registre.testnorge.arena.service.util.ServiceUtils;
-import no.nav.registre.testnorge.arena.service.util.TiltakUtils;
 import no.nav.registre.testnorge.arena.service.util.KodeMedSannsynlighet;
-import no.nav.registre.testnorge.domain.dto.arena.testnorge.brukere.Deltakerstatuser;
 import no.nav.registre.testnorge.domain.dto.arena.testnorge.tilleggsstoenad.Vedtaksperiode;
 import no.nav.registre.testnorge.domain.dto.arena.testnorge.vedtak.NyttVedtakResponse;
 import no.nav.registre.testnorge.domain.dto.arena.testnorge.vedtak.NyttVedtakTillegg;
@@ -31,14 +28,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import static no.nav.registre.testnorge.arena.consumer.rs.util.ConsumerUtils.UTFALL_JA;
-import static no.nav.registre.testnorge.arena.consumer.rs.util.ConsumerUtils.VEDTAK_TYPE_KODE_O;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyList;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.mockito.Mockito.times;
 
 @RunWith(MockitoJUnitRunner.class)
 public class RettighetTiltakServiceTest {
@@ -58,9 +50,6 @@ public class RettighetTiltakServiceTest {
     private IdenterUtils identerUtils;
 
     @Mock
-    private TiltakUtils tiltakUtils;
-
-    @Mock
     private ArbeidssoekerUtils arbeidssoekerUtils;
 
     @Mock
@@ -69,32 +58,15 @@ public class RettighetTiltakServiceTest {
     @InjectMocks
     private RettighetTiltakService rettighetTiltakService;
 
-    private Long avspillergruppeId = 1L;
-    private String miljoe = "t1";
-    private int antallNyeIdenter = 1;
-    private List<String> identer;
-    private NyttVedtakTiltak tiltakMedTilDatoFremITid;
-    private NyttVedtakTiltak tiltakMedTilDatoLikDagens;
-    private List<SyntRequest> syntRequest;
-
     @Before
     public void setUp() {
-        syntRequest = new ArrayList<>(Collections.singletonList(
-                SyntRequest.builder()
-                        .fraDato(LocalDate.now().toString())
-                        .tilDato(LocalDate.now().toString())
-                        .utfall(UTFALL_JA)
-                        .vedtakTypeKode(VEDTAK_TYPE_KODE_O)
-                        .vedtakDato(LocalDate.now().toString())
-                        .build()
-        ));
-        identer = new ArrayList<>(Collections.singletonList("01010101010"));
+        List<String> identer = new ArrayList<>(Collections.singletonList("01010101010"));
 
         Map<String, List<NyttVedtakResponse>> response = new HashMap<>();
         response.put(identer.get(0), Collections.singletonList(new NyttVedtakResponse()));
         response.get(identer.get(0)).get(0).setFeiledeRettigheter(new ArrayList<>());
 
-        tiltakMedTilDatoFremITid = NyttVedtakTiltak.builder()
+        NyttVedtakTiltak tiltakMedTilDatoFremITid = NyttVedtakTiltak.builder()
                 .tiltakId(123)
                 .tiltakAdminKode("IND")
                 .tiltakskarakteristikk("IND")
@@ -103,7 +75,7 @@ public class RettighetTiltakServiceTest {
         tiltakMedTilDatoFremITid.setFraDato(LocalDate.now().minusMonths(1));
         tiltakMedTilDatoFremITid.setTilDato(LocalDate.now().plusMonths(1));
 
-        tiltakMedTilDatoLikDagens = NyttVedtakTiltak.builder()
+        NyttVedtakTiltak tiltakMedTilDatoLikDagens = NyttVedtakTiltak.builder()
                 .tiltakId(123)
                 .tiltakAdminKode("IND")
                 .tiltakskarakteristikk("IND")
@@ -114,50 +86,6 @@ public class RettighetTiltakServiceTest {
 
     }
 
-
-    @Test
-    public void shouldOnlyOppretteTiltaksdeltakelseOgEndreDeltakerstatusGjennomfoeres() {
-        var vedtak = Collections.singletonList(tiltakMedTilDatoFremITid);
-        when(consumerUtils.createSyntRequest(antallNyeIdenter)).thenReturn(syntRequest);
-        when(tiltakSyntConsumer.opprettTiltaksdeltakelse(syntRequest)).thenReturn(vedtak);
-        when(identerUtils.getUtvalgteIdenter(avspillergruppeId, antallNyeIdenter, miljoe)).thenReturn(identer);
-        when(rettighetArenaForvalterConsumer.opprettRettighet(anyList())).thenReturn(new HashMap<>());
-        when(tiltakUtils.finnTiltak(identer.get(0), miljoe, tiltakMedTilDatoFremITid)).thenReturn(tiltakMedTilDatoFremITid);
-        when(tiltakUtils.canSetDeltakelseTilGjennomfoeres(tiltakMedTilDatoFremITid, Collections.singletonList(tiltakMedTilDatoFremITid))).thenReturn(true);
-        when(tiltakUtils.getVedtakForTiltaksdeltakelseRequest(tiltakMedTilDatoFremITid)).thenReturn(tiltakMedTilDatoFremITid);
-        when(tiltakUtils.getFoersteEndringerDeltakerstatus(anyString())).thenReturn(Collections.singletonList(Deltakerstatuser.GJENN.toString()));
-
-        rettighetTiltakService.opprettTiltaksdeltakelse(avspillergruppeId, miljoe, antallNyeIdenter);
-
-        verify(tiltakSyntConsumer).opprettTiltaksdeltakelse(syntRequest);
-        verify(identerUtils).getUtvalgteIdenter(avspillergruppeId, antallNyeIdenter, miljoe);
-        verify(rettighetArenaForvalterConsumer, times(2)).opprettRettighet(anyList());
-        verify(tiltakUtils).finnTiltak(identer.get(0), miljoe, tiltakMedTilDatoFremITid);
-        verify(tiltakUtils).getVedtakForTiltaksdeltakelseRequest(tiltakMedTilDatoFremITid);
-        verify(tiltakUtils).getFoersteEndringerDeltakerstatus(anyString());
-    }
-
-    @Test
-    public void shouldOppretteTiltaksdeltakelseOgEndreDeltakerstatusAvsluttet() {
-        var vedtak = Collections.singletonList(tiltakMedTilDatoLikDagens);
-        when(consumerUtils.createSyntRequest(antallNyeIdenter)).thenReturn(syntRequest);
-        when(tiltakSyntConsumer.opprettTiltaksdeltakelse(syntRequest)).thenReturn(vedtak);
-        when(identerUtils.getUtvalgteIdenter(avspillergruppeId, antallNyeIdenter, miljoe)).thenReturn(identer);
-        when(rettighetArenaForvalterConsumer.opprettRettighet(anyList())).thenReturn(new HashMap<>());
-        when(tiltakUtils.finnTiltak(identer.get(0), miljoe, tiltakMedTilDatoLikDagens)).thenReturn(tiltakMedTilDatoLikDagens);
-        when(tiltakUtils.canSetDeltakelseTilGjennomfoeres(tiltakMedTilDatoLikDagens, Collections.singletonList(tiltakMedTilDatoLikDagens))).thenReturn(true);
-        when(tiltakUtils.canSetDeltakelseTilFinished(tiltakMedTilDatoLikDagens, Collections.singletonList(tiltakMedTilDatoLikDagens))).thenReturn(true);
-        when(tiltakUtils.getVedtakForTiltaksdeltakelseRequest(tiltakMedTilDatoLikDagens)).thenReturn(tiltakMedTilDatoLikDagens);
-        when(tiltakUtils.getFoersteEndringerDeltakerstatus(anyString())).thenReturn(new ArrayList<>(Collections.singletonList(Deltakerstatuser.GJENN.toString())));
-        when(tiltakUtils.getAvsluttendeDeltakerstatus(tiltakMedTilDatoLikDagens, Collections.singletonList(tiltakMedTilDatoLikDagens))).thenReturn(Deltakerstatuser.FULLF);
-
-        rettighetTiltakService.opprettTiltaksdeltakelse(avspillergruppeId, miljoe, antallNyeIdenter);
-
-        verify(tiltakSyntConsumer).opprettTiltaksdeltakelse(syntRequest);
-        verify(identerUtils).getUtvalgteIdenter(avspillergruppeId, antallNyeIdenter, miljoe);
-        verify(rettighetArenaForvalterConsumer, times(2)).opprettRettighet(anyList());
-        verify(tiltakUtils).finnTiltak(identer.get(0), miljoe, tiltakMedTilDatoLikDagens);
-    }
 
     @Test
     public void shouldOppretteRettighetTiltaksaktivitetRequest() {
@@ -176,7 +104,7 @@ public class RettighetTiltakServiceTest {
 
         when(serviceUtils.velgKodeBasertPaaSannsynlighet(anyList())).thenReturn(aktivitetskodeMedSannsynlighet);
 
-        var request = rettighetTiltakService.opprettRettighetTiltaksaktivitetRequest("","",nyttVedtakTillegg);
+        var request = rettighetTiltakService.opprettRettighetTiltaksaktivitetRequest("", "", nyttVedtakTillegg);
 
         assertThat(request.getVedtakTiltak()).hasSize(1);
         assertThat(request.getVedtakTiltak().get(0).getFraDato()).isEqualTo(LocalDate.now().minusMonths(1));
